evententry/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-import-export==3.3.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
uritemplate==4.1.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
